+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### 0.1.0 (2021-08021)
+
+- 기능추가: on eventListener 추가(on 속성: text, callback)
+- 기능추가: constructor에 options parameter 추가 (polling, process)
+
 ### 0.0.5 (2021-08-20)
 
 수정 및 기능추가: